refactor(cli): split main.rs into cli/ subcommands and stac/root
